# docs/openapi.yaml
openapi: 3.0.3
info:
  title: DEEL BACKEND TASK - OpenAPI 3.0
  description: |-
    DEEL BACKEND TASK
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
  /contracts:
    get:
      tags:
        - contracts
      summary: List of contracts
      description: Returns a list of contracts belonging to a user (client or contractor), the list should only contain non terminated contracts.
      operationId: getContracts
      parameters:
        - in: header
          name: profile_id
          schema:
            type: string
          required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contract'
        '400':
          description: Client error
        '401':
          description: Not a valid profile
...
components:
  schemas:
...
    Contract:
      title: Contract response
      type: object
      properties:
        id:
          type: integer
          format: int64
        terms:
          type: string
        status:
          enum:
            - new
            - in_progress
            - terminated
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        ContractorId:
          type: integer
        ClientId:
          type: integer
...